Add files via upload
This is a skeleton script with the code to connect to ultrasound and some example widgets already written. 
To use this code open a downloaded or built version of 3D slicer, in the Modules: drop down menu select All modules and then finally Extension Wizard. 
Once in this module select Create Extension and enter your module name in Name:, keep the type as default and for the destination select your development repository. If you don't have a a development repository it is recommended to create a folder in  C: named d or dev to keep file paths short and select this folder. This is where the files will be generated. Then select Ok. 
Leave the Name: as is. If this is a trial module you can leave the category as Examples, if you know the category you wish to put your module in, for example IGT, you can enter that here. If you are unsure leave it as Examples as this can be changed later. Select Ok. 
Click on Add Module to Extension and name your module what ever you like, leave the Type: as scripted. Click OK. 
On the pop up block select Add selected module to search paths and click Yes. 
Go to your dev folder and select the folder you created, this will be the first Name: you entered. 
Within this folder will be another folder with your module name select this. Within this folder is a folder with the same name as your module and within this folder is a python script named YourModuleName.py open this and replace the contents of the script with this file.
Replace the "YourModuleName", "YourName" etc with the correct information, be sure to name your module the same name as the file. Once this is completed your module will be included in the lists of modules. 
To continue development include code to this file.